Editor++
designed a new editor, created mockups.
the design goal of this change was simply to be able
to handle more roomObjects in an easier way, based on
the discovery of many workflow problems, like adding
or modifying objects underneath widgets while still
trying to control their input.

the new approach is to move the object widgets into
their own screen, which can be shown based on a button
press, removing the various obj widget states,
simplifying that - plus input will only be processed
on the screen and could never affect objects in the
game world, which solves two problems. the redesign
also allows for two new obj widgets, for which we
don't have objects for yet - and it adds a nice 'notes'
section to let editors know there is a faster way
to work using shortcut keys along the top row of
a standard us keyboard.