The Colliseum - Part 13
improve the transition from buying a challenge
to starting a challenge. currently, the hero teleports
to the spawn point, with no fading in/out to black,
and it's really jarring and seems buggy.

after some hacking on the game, a few things were
discovered: if you try to add a dialog screen when
the vendor screen is closing, the game freezes and
no input is processed at all. hilarious! if you
try to add a dialog screen at the end of the
BeginChallenge() routine, the same thing happens.
amazing! glorious! plus, there's no transition.

the better solution was to call BuildLevel()
at the start of BeginChallenge(), passing in the
current level's ID. this causes the level screen's
overlay rec to be set to 100% opacity, and then
fades out, smoothly displaying the game. this is better
transition than previous teleport transition.
smooth, like a saxophone submerged in butter.