SeekerExploder++
migrate seekerExploder actor to be a roomObject.

Why are we doing this?
Seekers are implemented as actors, of which there
can only be 30 on screen at once, due to the large
amount of checks each actor goes through. they are
expensive to process, + have a ton of animation
frames too. well, the seeker can be modeled as a
roomObj instead, with all the same behavior, but
we also gain: the roomObj count max is 3000, so we
can have A TON MORE seekers on screen, drastically
increasing the risk/skill level. plus, it cuts
processing time down for the actor pool,
but this is negated by the additional number of
seekers that will likely exist on screen.

basically, instead of having 1 actor seeker, we
can now have like 10 to 50 for about the same cost.

had to hack around the interaction functions to make
a non-blocking roomObject interact with projectiles
as a blocking roomObject - which was fine.

rewrote how seekers die, done thru HandleCommon()
now - if the seeker is hit by a projectile that
could kill it, or if the seeker gets close enough
to the hero, HandleCommon() is called, at which point
the seeker creates an explosion and drops loot.
so, we can still cause explosion chains with seekers,
which was kinda what made them awesome - they're
like exploding barrels that move around and chase you.

rewrote how seekers explode - when they get near
hero they instantly become an explosion, making them
a bit more dangerous.

improved seekers speed - moves slowly when they
cant see hero, faster when they can see him.

created exploding object gameobject.
this conceptually represents any object that should
explode. anyobject can become an exploding object.
the exploding object's definition doesn't modify
anything about the previous gameobject, it simply
sets getAi to true, and reset the lifecounter,
then sets the lifetime. once the life counter
hits the lifetime, the object spawns an explosion
and is released, spawning some loot too.

using this simple approach, barrels and seekerExploders
are now both being fed into the exploding object
system, and working as expected. this removed
exploding barrels projectile, and prevented me
from adding exploding exploder roomobjects.

because seekers arent' blocking, they can walk thru objects.
so, when a seeker interacts with a blocking roomObject, they
now explode. this change deals the boss a bit more damage
overtime, because the seekers run into walls alot.

as of now, seekers are implemented well enough.
their only downside is that they kill their boss
if enough of them are spawned in a room. but,
they would also kill the hero too - so i see this
as balanced for now. if hero can survive long enough,
the boss's mobs will kill him.

updated the uwp project's content folder.
set game into release mode for this commit as well.