Wall Jumps
the basic idea is that we want the player
to be able to detach from the wall (B button
press), starting a fall. Then using the
joystick or Dpad, changing the direction
of the falling hero overtime. Once the
player realizes this is possible, they will
eventually line up a path that intersects
a ladder or foothold, and try to grab it
via the A button press - completing a heroic
leap from one spot to another.

extended falling routine in actor.update to
allow for setting the hero's interaction rec
while in statelocked falling state. this allows
player to spam A button while falling to
heroically grab at any nearby ladder or foothold.
-made sure to get the current controller's
direction just prior to checking the grab,
so it doesn't feel laggy or a frame behind.

Done
so, that *completely* implements wall jumps.
playtested, wall jumping is incredibly fun.
this allows for climbing up high to fall
to catch a foothold in an unreachable area,
all based on skill and understanding of the game.

added soundfx and attention particle when hero
successfully completes a wall jump. we could
also add any kind of celebration here to note
the heroic leap, and we could also keep track
of how many successful wall jumps hero has made.

added wallJumps int to playerData,
linked it up to wall jump rountine,
which now increments the wallJump int
each time the hero completes a wall jump.
lets use this data to unlock
a challenge or trophy or something...

Achievements
created achievements enumerator,
with wallJumps as it's only value for now.
created fun_hero.CheckAchievements(),
which accepts an achievement enum,
which then determines which achievement
routine to run. two are implemented
currently - checking for 10 and 100
successful wall jump - from playerData.

successfully completing an achievement
pops a dialog screen telling player
what they have completed. in the future,
this WILL lead to more meaningful
effects, because a dialog screen feels
cheap, honestly. playtested, working
as expected.

updated UWP to match Dx.