Actor Widgets
created widgetActors.cs to contain actors,
for editor development testing purposes, based
on the concept and design of object widgets.
for now, this widget simply displays which enemies
belong to which dungeons, but in the near future
actors will be selectable and add-able to the
game world from editor, via this actor widget.

implemented forest actors widget. this displays
the boss, miniboss, and standard actors required
to create a dungeon.

moved level obj widgets to the right of dungeon objs
in editor menu screen to mek room for actors widget.

when the level type button changes, the
corresponding actor widget is displayed. for now,
only the forest actors widget is display.

added town actors widget, with blob as only actor.
playtested. updated uwp project too.