Editor++
we want to be able to add actors to dev levels
in editor, to test their behaviors. the actor
widget exists, we just need to implement the
logic for 'adding' a selected actor.

implemented adding actors via editor menu
screen. this is just another 'state' of the
widget object tools class, creating an actor
instead of a gameobject upon clicking in world.
we also display some info about the actor
in the objTools widget, which we have completely
snatched from conceptually just managing
objects to managing both actors and objects.