Climbing with pets
pets should match hero's position, plus 4 or 6 pixels
on y axis, if hero is climbing or falling. this means
pet teleports to hero's back upon a climb. while pet is
on hero's back, pet ignores all collisions and their ai
changes to simply play the pet's soundfx randomly.
done, and basically working.

updated initial falling routine for climbable walls
to ignore hero's pet if hero is climbing with pet.
this prevents the stuttering effect and the initial
push particles from being created when hero starts
climbing the wall, because the pet object was being
forcibly teleported to the hero's position - then
the walls were trying to push it and thought it
was the objects initial fall too.

added code to sort pet over hero while in climbing
state. this helps to sell the effect of having the
pet in the hero's backpack.

fixed a bug where pets would display their swimming
animation on the wall, if hero began his climb from
a swimming state (in water).

fixed a bug where pets would fly off the edge of a
cliff that hero had just climbed. because we're
teleporting pet onto hero's back, pet is still
overlapping a wall object when hero completes his
climb - causing the pet to instantly fall.

updated mountain level, fixing an annoying issue
with the easiest path for climbing up. a bush was
in the way, preventing a proper landing state from
triggering.

updated UWP branch with latest content files.