Infinite Fairies Cheat
this allows for damage and death to occur, but prevents
a gameover. add a boolean into flags, which is checked in
bottle use routines. if true, the upon heros death game
assume a fairy bottle is present. add this cheat to the
cheats menu. done. tested. working.

this is actually fun to play around with too.