Climbing++
when hero first attaches to a foothold (if hero's
state isn't climbing), we should align hero to the
base of the foothold (but only on the Y axis). why?
hero's sprite's center position MUST overlap a foothold
obj in order for hero to stay in a climbing state.
by placing the hero at the obj's position, plus an
offset (for taste), we can ensure hero properly
attaches to wall when the player simply presses A.

this prevents a bug where if player isn't pressing
up on dpad (when they press A to begin the climb)
then hero just falls off wall! - because
the center of hero's sprite isn't touching the
foothold object YET. if we align these two initially,
we prevent this state from occurring, and player
can simply press A to attach to wall. we dont
align the X position because we don't want to
affect the hero's position too much (jarring),
plus we already know the interaction point has
collided with the foothold (so the hero is 'close
enough' to grab onto the foothold obj).