Polish++
if a tree is burning, we can't destroy it.
because we don't check for burning trees in BlowUp().
now we do.

we should add an offset to the camera when it
centers to the dungeon room - it needs a +16 offset
on the Y axis in order to properly display large
rooms like columns.

Refactoring
functions_level.buildPosition should be level.buildPosition.
same with fun_level.dungeonTrack... lol. < done.