I have no idea what I'm doing
*actually* tested deployment on win7 and of course,
it doesn't work. due to some weird publishing settings.
i can't actually get the game installed on win7, but
i fixed the publishing settings so the game can be
installed on win10 just fine. which means all previous
versions likely didn't install on windows at all. oops.
which means the only way to play the game was via uwp.
i'm sorry. i'm so, so sorry. that's embarassing.

it has something to do with the manifest file and
the click-once deployment that prevents the game
from being installed on windows 7. but fear not,
if you clone the repo and build it in monogame on
win7 - it will run just fine. so, learn to fish.

updated readme.