JIT Fix
so when testing on the xbox one, and on the pc, I
noticed that there was a pause before the first build
of any dungeon. subsequent builds were very fast,
leading me to believe that there is an initial setup
cost or a just-in-time compilation happening on the
first build. the fix is simply to build a forest dungeon
when the game first boots up, incurring this cost when
the game isn't doing anything important, like running.
thanks .Net!

fix implemented, but it exposes a flaw in the design of
Functions_Level, which relys on a level screen being
present (via a reference). so had to remove some code
there to make this boot build fix work properly.
there are plans to make screens global later on that
will resolve this problem, so i'm not fixing the
minor visual bug it creates: level screens sometimes
don't fade in from black properly now.