Arrow Count++
issue: the arrow count (in inventory widget) can sometimes
be displayed as 0, when hero doesn't have bow - leading to a
state where the menuItem is unknown and has a count of 0,
which visually makes NO SENSE at all.

fix inventory widget to only display arrow counter if
hero actually has bow boolean set to true. this is only
checked when the widget or scrveen is initially opened.
not when we reset/setup the inventory menuItems.

updated SetInventoryMenuItems() to set visible booleans
properly for bomb and arrow counts. bomb count is always
displayed, while arrows are dependent upon the bow boolean.
tested, working.