Exiting Dungeons
this should return the hero to the dungeon entrance,
on the proper dungeon entrance level. hero should spawn
just south of dungeon entrance. this means dungeon entrances
can't be blocked by anything directly infront of them.
to do this, we'll need to drastically modify how spawning
is done, via spawnPos.

SpawnPos++

Removed spawnPos from rooms and put into Functions_Hero,
for various reasons explained below.. this was done in
prep for exiting dungeons to their corresponding field
levels and spawning in front of the dungeon entrance object.

what do we gain by storing spawnPos in room?
actually nothing, because the data isn't used in any special
way related to the lifecycle of rooms. in fact, it would be more
useful and less wasteful if we only had one spawnPos instance,
because we have spawnPos instances for each room in the level,
which currently can upwards of 100 rooms. so we have 100 extra
spawnPos references sitting in memory, not really being used
except when the hero enters that specifc room. that's dumb.
simplified room class.

upon implementation, it became apparent that spawnPos should
not live in functions_hero, because we actually need two
position values, not one, and they map better to the
field and dungeon concepts designed in LevelSet class.

created two spawnPos vector2 instances in levelSet, one
for field and one for dungeon. this seperates them and
places them in the codebase in a logical area (levelSet).

upgraded colliseum pit to exit to colliseum exterior.
this also sets up the foundation for linking levels to other
levels (fields) using entrance objects.

+ fixed issue with dungeon exit obj not properly aligning.
a ton of playtesting to ensure entering and exiting various
entrances and levels works properly with the new spawn system.