Beefy Bats
* beefy bats dont properly face their moving direction.
their sprite facing direction is independent of their moving direction.
- this was a problem with their animation frames, which
was fixed. however, upon interacting with them a new bug was
discovered with beefy bats: since all ai controlled actors
move diagonally, the diagonal movement is resolved to a cardinal
direction and all diagonal movement specifically resolves to
either left or right (not up or down). as a result, the beefy
bat sprite rarely displays his up or down animation - unless
he bite attacks or dashes north / south. this isn't the visual
process that I want. instead, we'll update the AI to set the
movement to a cardinal direction for beefy bats, instead of a
diagonal direction - allowing the visual process to occur
the way I want it to.

^ done. it would appear a better solution would be to model
if an actor chases diagonally or cardinally using a boolean.
the current system specifically checks the actor.type, of
which there are some and will be many more. it would save time
and code later, but incurs an upfront cost of a quick system
rewrite for actor ai chasing... seems like a good investment.

* instead of checking by actor type, we want to add a boolean
to actor class that controls how this actor chases or moves towards
the hero (diagonally or cardinally). moving diagonally is preferred,
but there are certain actors that require north/south movement
to properly display their animated sprites.. moving cardinally
was the original implementation for actors, until it was realized
that actors easily get stuck on blocks and never move around them
until hero does, leading a variety of exploits that made the ai
look really, really dumb. the fix was simply to make all actors
chase using diagonal movement, which allowed them to slide around
blocks (thanks physics system!), making them very challenging and
"smart".

^ done. notes from playtesting: we should only be spawning one
beefy bat in dungeon rooms for spawn points. this may of already
been implemented - but anyway - we also need to increase how
often the beefy bat dashes towards hero - and increase his attack
probability to 100% - but we also need to decrease his visibility
radius to hero so he doesn't trigger until hero is much closer.
at that point, the bat is a very dangerous enemy.

to actually implement this without complicating the AI too much,
I decided to create an new ActorAI enum value for BeefyBat.
this will be a seperate ai path, much like miniboss or boss.

beefybat has a 90% chance to bite attack hero, instead of the
standard 75% chance. the new AI also returns every bat to a south
facing state after an idle dash event, so a hivemaind visual
is conveyed (they all face the same way, watching you). it's
creepy and effective and exactly what I want.

* bats should not be able to die from bite attacks, because it's
very likely other bats will be group attacking and kill each other
in this case. < done. this also improved the structure of Damage(),
creating sections for standard, miniboss, and boss regions.

the playtesting of beefybats has been successful - they now
have a unique and interesting 'personality' as an enemy type.