Fall Sfx++
we dont have fall sfx for hero, which feels missing.

hacked into interactions between mid mountain wall and
actor, the terminal velocity check was moved to later
in the method, then an if branch was added prior,
checking to see if the actor's falling magnitude (Y axis)
was less than the terminal velocity. if it is, then
the initial fall sound fx is played.

notes: it only takes a few frames for the actor to reach
terminal velocity from falling, so for those few frames
the fall soundfx is called each frame. a sound effect
cannot be played again until it's finished playing, so
the additional calls to play the sound effect are essentially
ignored and initial falling sound is implemented again.