Swamp Miniboss - OctoMouth
this is lttp's darkworld walking mouth migrated to the
water, representing an underwater creature with giant
teeth that shoots fireballs and can swim, dive.

underwater, octomouth randomly moves in a direction or
chases hero if he's low on health. often, octomouth will
rise up out of the water and stop all his movement.
then he will face towards the hero and wait to be attacked.
sometimes he will shoot a fireball at hero, and sometimes
he will dive back underwater. if his health gets below
half, he moves alot faster. that's it.

Improvements

+ if any actor is underwater, they cannot take any damage
from anything and are essentially invincible. the hero can
go underwater, but it only lasts for 4 seconds, and while
underwater the hero can't attack or use items. it's defense.

+ this allows various water-based enemies to retreat into the
water as part of their ai strategy. these enemies then move to
a different location to attack from a distance or harass upclose.

+ dashing while underwater is now allowed. hero can dash, along
with enemies. this allows underwater swamp enemies more freedom
to move about their surroundings.

+ underwater enemies that are hit immediately go into their
underwater state - so hero only gets one hit at a time. else,
it's very easy to spam attack and that feels cheap - esp for
these underwater enemies which could easily dive afterwards.

+ modified octomouth mouth to more often surface above water,
and reduced his affinity to dive to be more rare. this allows
the player more time to attack him when he's above water, and
decreases the overall time he's underwater so hero gets more
chances to attack him. this is to balance the effect of hitting
underwater enemies, which is that they go underwater immediately.
so the attack window is open more often, but it shuts as soon
as an attack lands and the player has to wait for a moment for
the next window. this is noted with auditory soundfx (each event),
so the attack window and subsequent failure/success feels good.

+ reduced the health of octomouth from 15 to 10, since it takes
more time and patience to kill this particular miniboss and i
don't want the pacing of battle to slow down too much for this
1v1.

+ updated actor animation routines to allow swimming underwater
enemies to have an attack animation, since octomouth is always
swimming and MUST display some visual change when he attacks.

+ updated dungeon routines to spawn octomouth in swamp dungeon.
while there isn't water in the hub room, octomouth is setup to
function as if there were, and 'dives underwater' into the ground.
which works system-wise (he'll pop back up 'above water'), but
doesn't make much sense visually..