Swamp Boss : Octohead + Tentacles
tentacles are actors with 30 health.
upon getting hit, they duck underwater for a time,
then surface again to slowly chase hero (no dashing).
this is because there will likely be more than 1 on
screen at a time, so this is a group strategy.

octohead simply ducks in and out of water and
randomly moves around. he has 8 health, and
spawns a tentacle each time he comes up out of the
water. this means he spawns as many tentacles as
he has HPs. this also lets the player choose how
they want to deal with the tentacles: let them
stackup or deal with them one at a time.

upon defeat of the head, all tentacles explode and die.

+++ improved SetActorInput() to base the ai's distance off
the enemy hitBox's center vs the hero's hitBox's center.
previously, this was based off the actor's sprite's center.
but, the hitBox it what we should of been using all along,
because the proceeding attacking routines are setup based
upon that assumption. so this change makes enemy attacks
much more accurate, because it removes an offset to the
Y axis for enemies larger than 16 pixels tall. standard
actors still function as normal, with slightly better
accuracy and without a bias for attacking north.

+++ updated Projectile.Spawn() to take an Actor as a
parameter, setting up this method to handle a big change
coming up soon. previously, this was taking a move
component, which didn't provide the hitBox info needed
to *actually* align projectiles to their casters.
now this method has that information, but the projectile
system isn't setup to work with that info or store it
frame to frame (that info only exists inside the method
when it's called - consider the scope).

+ boomerang shouldn't bother / hurt octohead, otherwise he
spawns a tentacle without taking damage. done.