Boat++
added back of boat. remade boat level so
boat is further away from bottom of level.
this makes it easier to explore all around
the boat.

changed water tiles to sort to -256 on zxis.
they just sort really far behind eveything now.
this fixes sprite issue with back of boat.

added center part of back boat. lots of testing
to make sure proper collisions with all of
boat in and out of water.