Overworld++
swamp dungeon/entrance no longer exists on island.
this level now only exists on sea overworld.

created Functions_overworld.OpenMap(), which acts as
a controller to determine which overworld map should
be loaded, based on the last level hero was on.

updated codebase to load overworld screens thru
overworld.openMap(). however, this cannot be tested
as is because we don't have a way to physically get
to the sea overworld map yet - this will be done via
Brandy, the Ship's Captain.