Ship Travel
brandy is now a vendor.
added vendor dialog for brandy.
added ship destinations to menuItem enum.
added ship destination defs to menuItem.setType().
added routines to PurchaseItem() to complete the
functionality of destination purchases. the level
is rebuilt based on a new levelID value, which is
determined by the type of menuItem destination
purchased. we can now travel between overworlds.
tested and working as expected.