Ship Travel++
purchasing a trip should exit to the overworld
screen, but should also set the current level.
that way, the overworld screen uses the current
level value to determine which overworld to display.
this has the net effect of displaying the proper
overworld map when hero purchases a trip from
boat captain, brandy.