Summary Screen++
summary screen is never reset.
updated summary screen to set it's display state to
opening via Open() method. this allows summary screen
to display more than once. duh. ouch.

summary screen sets the current level based on saveData.
this used to be useful, but now we always know what the
last level loaded was, via levelSet.currentLevel. so, we
dont really need to set the level from summary screen.
this has the net effect of allowing the player to die
in the swamp and then displays the sea overworld,
instead of the island overworld - keeping things consistent.

tested, working. dying in a level or dungeon
returns link to the proper overworld screen,
once the summary screen has completed.